Consitently use ⏩ and ✅ throughout the install process

+++ lib/install/config.rb
@@ -7,15 +7,22 @@ return if pack_path_missing?
 step_path = "/app/javascript/config/"
 index_src = fetch(step_path, "index.js.tt")
 index_path = config_path / "index.js"
+friendly_index_path = index_path.relative_path_from(Rails.root).to_s
 stimulus_reflex_src = fetch(step_path, "stimulus_reflex.js.tt")
 stimulus_reflex_path = config_path / "stimulus_reflex.js"
+friendly_stimulus_reflex_path = stimulus_reflex_path.relative_path_from(Rails.root).to_s
 cable_ready_src = fetch(step_path, "cable_ready.js.tt")
 cable_ready_path = config_path / "cable_ready.js"
 
 empty_directory config_path unless config_path.exist?
 
-backup(index_path, delete: true) do
-  copy_file(index_src, index_path)
+if index_path.exist?
+  say "⏩ #{friendly_index_path} already exists. Skipping"
+else
+  backup(index_path, delete: true) do
+    copy_file(index_src, index_path)
+  end
+  say "✅ Created #{friendly_index_path}"
 end
 
 index_pattern = /import ['"](\.\.\/|\.\/)?config['"]/
@@ -28,14 +35,19 @@ if pack.match?(index_pattern)
     matches = lines.select { |line| line =~ index_commented_pattern }
     lines[lines.index(matches.last).to_i] = index_import
     pack_path.write lines.join
+
+    say "✅ Uncommented StimulusReflex and CableReady configs imports in #{friendly_pack_path}"
+  else
+    say "⏩ StimulusReflex and CableReady configs are already being imported in #{friendly_pack_path}. Skipping"
   end
 else
   lines = pack_path.readlines
   matches = lines.select { |line| line =~ /^import / }
   lines.insert lines.index(matches.last).to_i + 1, index_import
   pack_path.write lines.join
+
+  say "3✅ StimulusReflex and CableReady configs will be imported in #{friendly_pack_path}"
 end
-say "✅ SR/CR configs will be imported in #{friendly_pack_path}"
 
 # create entrypoint/config/cable_ready.js and make sure it's imported in application.js
 copy_file(cable_ready_src, cable_ready_path) unless cable_ready_path.exist?
@@ -43,31 +55,33 @@ copy_file(cable_ready_src, cable_ready_path) unless cable_ready_path.exist?
 # create entrypoint/config/stimulus_reflex.js and make sure it's imported in application.js
 copy_file(stimulus_reflex_src, stimulus_reflex_path) unless stimulus_reflex_path.exist?
 
-if ["webpacker", "shakapacker"].include?(bundler)
-  append_file(stimulus_reflex_path, <<~JS, verbose: false) unless stimulus_reflex_path.read.include?("StimulusReflex.debug")
-
-    if (process.env.RAILS_ENV === 'development') {
-      StimulusReflex.debug = true
-      window.reflexes = StimulusReflex.reflexes
-    }
-  JS
-elsif bundler == "vite"
-  append_file(stimulus_reflex_path, <<~JS, verbose: false) unless stimulus_reflex_path.read.include?("StimulusReflex.debug")
-
-    if (import.meta.env.MODE === "development") {
-      StimulusReflex.debug = true
-      window.reflexes = StimulusReflex.reflexes
-    }
-  JS
+if stimulus_reflex_path.read.include?("StimulusReflex.debug =")
+  say "⏩ Development environment options are already set in #{friendly_stimulus_reflex_path}. Skipping"
 else
-  append_file(stimulus_reflex_path, <<~JS, verbose: false) unless stimulus_reflex_path.read.include?("StimulusReflex.debug")
+  if ["webpacker", "shakapacker"].include?(bundler)
+    append_file(stimulus_reflex_path, <<~JS, verbose: false)
 
-    // consider removing these options in production
-    StimulusReflex.debug = true
-    window.reflexes = StimulusReflex.reflexes
-    // end remove
-  JS
+      if (process.env.RAILS_ENV === 'development') {
+        StimulusReflex.debug = true
+      }
+    JS
+  elsif bundler == "vite"
+    append_file(stimulus_reflex_path, <<~JS, verbose: false) unless stimulus_reflex_path.read.include?("StimulusReflex.debug")
+
+      if (import.meta.env.MODE === "development") {
+        StimulusReflex.debug = true
+      }
+    JS
+  else
+    append_file(stimulus_reflex_path, <<~JS, verbose: false)
+
+      // consider removing these options in production
+      StimulusReflex.debug = true
+      // end remove
+    JS
+  end
+
+  say "✅ Set useful development environment options in #{friendly_stimulus_reflex_path}"
 end
-say "✅ Set useful development environment options"
 
 complete_step :config
